requirements.txt
feedparser == 5.1.3
beanstalkc == 0.3.0
PyYAML == 3.10
